+++ junit/kubernetes-server-mock/src/main/java/io/fabric8/kubernetes/client/server/mock/crud/PutHandler.java
@@ -16,6 +16,7 @@
 package io.fabric8.kubernetes.client.server.mock.crud;
 
 import com.fasterxml.jackson.databind.JsonNode;
+import io.fabric8.kubernetes.api.model.GenericKubernetesResource;
 import io.fabric8.kubernetes.client.utils.Serialization;
 import io.fabric8.mockwebserver.crud.AttributeSet;
 import okhttp3.mockwebserver.MockResponse;
@@ -36,12 +37,18 @@ public class PutHandler implements KubernetesCrudDispatcherHandler {
 
   @Override
   public MockResponse handle(String path, String contentType, String requestBody) throws KubernetesCrudDispatcherException {
-    validateRequestBody(requestBody);
+    GenericKubernetesResource resource = validateRequestBody(requestBody);
     final AttributeSet attributes = persistence.getKey(path);
 
     final Map.Entry<AttributeSet, String> currentResourceEntry = persistence.findResource(attributes);
     if (currentResourceEntry == null) {
       return new MockResponse().setResponseCode(HttpURLConnection.HTTP_NOT_FOUND);
+    }
+
+    // Delete the resource if it is marked for deletion and has no finalizers.
+    if (resource.isMarkedForDeletion() && resource.getFinalizers().isEmpty()) {
+      persistence.processEvent(path, attributes, currentResourceEntry.getKey(), null);
+      return new MockResponse().setResponseCode(HttpURLConnection.HTTP_OK);
     }
 
     final JsonNode currentResource = persistence.asNode(currentResourceEntry);